+++ editor_patch/event.h
@@ -51,7 +51,9 @@ enum class AlpineDedEventID : int
     Set_Debris,
     Set_Fog_Color,
     Set_Entity_Flag,
-    AF_Teleport_Player
+    AF_Teleport_Player,
+    Set_Item_Drop,
+    AF_Heal
 };
 
 constexpr int af_ded_event_to_int(AlpineDedEventID event_id) noexcept