+++ game_patch/rf/ai.h
@@ -220,6 +220,15 @@ namespace rf
     static auto& ai_get_attack_range = addr_as_ref<float(AiInfo& ai)>(0x004077A0);
     static auto& ai_has_weapon = addr_as_ref<bool(AiInfo *ai_info, int weapon_type)>(0x00403250);
     static auto& ai_add_weapon = addr_as_ref<void(AiInfo *ai_info, int weapon_type, int clip_count)>(0x004030D0);
+    static auto& ai_set_target = addr_as_ref<void(AiInfo* ai_info, int target_handle)>(0x00409050);
+    static auto& ai_set_mode =
+        addr_as_ref<void(AiInfo* ai_info, AiMode mode, int mode_param_0, int mode_param_1)>(0x00407E20);
+    static auto& ai_set_submode =
+        addr_as_ref<void(AiInfo* ai_info, AiSubmode submode)>(0x00407E80);
+    static auto& ai_enter_attack_mode = addr_as_ref<void(AiInfo* ai_info)>(0x00408AC0);
+    static auto& ai_acquire_new_target = addr_as_ref<void(AiInfo* ai_info)>(0x00403A80);
+    static auto& ai_path_locate_pos = addr_as_ref<void(Vector3* pos, float radius, float height, bool walk_only,
+        GPathNode** first_node, GPathNode** second_node, bool a7)>(0x0040C2C0);
 
     static auto& ai_pause = addr_as_ref<bool>(0x005AF46D);
 }