+++ game_patch/rf/object.h
@@ -177,6 +177,8 @@ namespace rf
     static auto& obj_light_calculate = addr_as_ref<void()>(0x0048B0E0);
     static auto& physics_force_to_ground = addr_as_ref<void(Object* obj)>(0x004A0770);
 
+    static auto& obj_set_friendliness = addr_as_ref<void(Object* obj, int friendliness)>(0x00489F70);
+
     static auto& object_list = addr_as_ref<Object>(0x0073D880);
 
     static auto& debris_create = addr_as_ref<Debris*(int parent_handle, const char* vmesh_filename,